move colorutils, vectors and threadpool into model
