+++ fract4d/c/model/colorutils.cpp
